# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/ottosfa/v2.3/todolist/bulk": {
            "post": {
                "description": "Todolist - Upload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todolist"
                ],
                "summary": "Todolist - Upload",
                "operationId": "Todolist - Upload V2.3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Body with file csv",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolist - Upload EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/todolist/update": {
            "post": {
                "description": "Todolist - Update V2.3",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todolist V2.3"
                ],
                "summary": "Todolist - Update V2.3",
                "operationId": "Todolist - Update V2.3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTodolistV2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Update EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2.3/village/list": {
            "get": {
                "description": "Village - List by district",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Village V2.3"
                ],
                "summary": "Village - List by district",
                "operationId": "Village - List by district",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Create EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/institution/list": {
            "post": {
                "description": "Institution - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Institution - List",
                "operationId": "Institution - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/merchant-type/list": {
            "get": {
                "description": "MerchantType - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "MerchantType - List",
                "operationId": "MerchantType - List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/province/list": {
            "post": {
                "description": "Province - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Province - List",
                "operationId": "Province - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Province - List Response EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dbmodels.Provinces"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/todolist/bulk": {
            "post": {
                "description": "Todolist - Upload",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todolist"
                ],
                "summary": "Todolist - Upload",
                "operationId": "Todolist - Upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Body with file csv",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolist - Upload EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/ottosfa/v2/todolist/update": {
            "post": {
                "description": "Todolist - Update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Todolist"
                ],
                "summary": "Todolist - Update",
                "operationId": "Todolist - Update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 7n1cdTNMK_EeoXr_0D0luYJ68NaYESWxyPXd",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTodolist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todolists - Update EXAMPLE",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/ottosfa/v2/village/list": {
            "post": {
                "description": "Village - List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "List"
                ],
                "summary": "Village - List",
                "operationId": "Village - List",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Village - List Response EXAMPLE",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dbmodels.Village"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dbmodels.Provinces": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "IDM"
                },
                "province_code": {
                    "type": "string",
                    "example": "IDM"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dbmodels.Village": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "district_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "village_code": {
                    "type": "string"
                }
            }
        },
...
        "models.Meta": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "OK"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/models.Meta"
                }
            }
        },
...
        "models.SearchReq": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string",
                    "example": "IDM"
                }
            }
        },
        "models.TaskParams": {
            "type": "object",
            "properties": {
                "file_edukasi": {
                    "type": "string"
                },
                "sub_category_id": {
                    "type": "string"
                },
                "supplier_name": {
                    "type": "string"
                }
            }
        },
        "models.UpdateTodolist": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "merchant_id": {
                    "description": "ID of merchant or merchant_new_recruitment",
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "sales_phone": {
                    "type": "string"
                },
                "task_attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TaskParams"
                    }
                },
                "task_date": {
                    "type": "string"
                }
            }
        },
        "models.UpdateTodolistV2": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "AIA Central Lt. 27, Jl. Jend. Sudirman No.Kav 48A, RT.5/RW.4, Karet Semanggi, Setia Budi, RT.5/RW.4, Karet Semanggi, Kota Jakarta Selatan, Daerah Khusus Ibukota Jakarta 12930, Indonesia"
                },
                "address_benchmark": {
                    "type": "string",
                    "example": "sebelah rumah sakit jiwa"
                },
                "category_id": {
                    "type": "string",
                    "example": "6"
                },
                "id": {
                    "type": "string",
                    "example": "136129"
                },
                "merchant_id": {
                    "description": "ID of merchant or merchant_new_recruitment",
                    "type": "string",
                    "example": "2"
                },
                "merchant_phone": {
                    "type": "string",
                    "example": "082222230118"
                },
                "notes": {
                    "type": "string"
                },
                "owner_name": {
                    "type": "string",
                    "example": "Wawan"
                },
                "sales_phone": {
                    "type": "string",
                    "example": "082113776997"
                },
                "sales_type_id": {
                    "type": "integer"
                },
                "task_attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TaskParams"
                    }
                },
                "task_date": {
                    "type": "string",
                    "example": "2021-08-24"
                }
            }
        }
    }
}